Fixed events clogging up the SDL poll queue
First, the idris-sdl2 function for polling had to be modified so as to
poll until there were no more events left. Doing otherwise meant the event
queue got clogged up, being consumed at an event/frame rate.

The second issue was in returning `Nothing` in case of `SDL_MOUSEMOTION`
events (this is part of the C code that packages events for Idris),
causing the aforementioned Idris polling function to stop consuming
events for the current frame, always effectively returning a single event.